Move auth token to the entities subdirectory directory
