+++ Questrade/ISignInService.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 using ParityService.Questrade;
-using ParityService.Questrade.Models;
+using ParityService.Questrade.Models.Entities;
 
 namespace ParityService.Questrade
 {
